requirements.txt
coloredlogs
tabulate
python-dateutil
PyYAML>=5.0
rotkehlchen @ git+https://github.com/coinyon/rotki.git@bitcoinde#egg=rotkehlchen
jinja2
